Search_movies/client_server+sql/requirements.txt
mysql-connector-python~-8.0.28